backend/requirements.txt
anthropic==0.79.0
beautifulsoup4==4.14.3
fastapi==0.128.8
html2text==2025.4.15
langchain_anthropic==1.3.3
langchain_chroma==1.1.0
langchain_core==1.2.11
langchain_huggingface==1.2.0
langchain_text_splitters==1.1.0
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
slowapi==0.1.9
uvicorn==0.40.0
